requirements.txt
# Core FastAPI and server requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# HTTP requests and file handling
requests>=2.31.0
httpx>=0.25.0

# PDF processing
PyMuPDF>=1.23.0

# Machine Learning and Search
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.30.0

# LLM API clients
openai>=1.0.0
cohere>=4.0.0

# Utility packages
tqdm>=4.65.0
logging>=0.4.9.6
typing-extensions>=4.5.0

# Optional: For better JSON handling
ujson>=5.8.0

# Optional: For async operations
aiofiles>=23.0.0
asyncio>=3.4.3

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0